Add files via upload
Steganography is the art of hiding the fact that communication is taking place, by hiding information in other information.
This project hides the message with in the image, text file, audio file and video file. In this project, the sender selects a cover file (image, text, audio or video) with secret text and hide it into the cover file by using different efficient algorithm and generate a stego file of same format as our cover file (image, text, audio or video). Then the stego file is sent to the destination with the help of private or public communication networks. On the other side i.e. receiver, the receiver downloads the stego file and by using the appropriate decoding algorithm retrieves the secret text that is hidden in the stego file.



CODE ADAPTED AND MODIFIED 

ADAPTED FROM - https://github.com/Priyansh-15/Steganography-Tools/blob/master/Steganography.py